+++ tests/insights/test_shap.py
@@ -7,6 +7,7 @@ from unittest import mock
 import numpy as np
 import pandas as pd
 import pytest
+from matplotlib import pyplot as plt
 from pytest import mark
 
 from baybe._optional.info import INSIGHTS_INSTALLED
@@ -139,11 +140,7 @@ def test_invalid_explained_data(ongoing_campaign, explainer_cls, use_comp_rep):
         use_comp_rep=use_comp_rep,
     )
     df = pd.DataFrame({"Num_disc_1": [0, 2]})
-    with pytest.raises(
-        ValueError,
-        match="The provided dataframe must have the same column names as used by "
-        "the explainer object.",
-    ):
+    with pytest.raises(ValueError, match="must contain all columns that were used"):
         shap_insight.explain(df)
 
 
@@ -159,8 +156,10 @@ def test_plots(ongoing_campaign: Campaign, use_comp_rep, plot_type):
     df = ongoing_campaign.measurements[[p.name for p in ongoing_campaign.parameters]]
     if use_comp_rep:
         df = ongoing_campaign.searchspace.transform(df)
+
     with mock.patch("matplotlib.pyplot.show"):
         shap_insight.plot(plot_type, df)
+        plt.close()
 
 
 def test_updated_campaign_explanations(campaign, n_iterations, batch_size):