+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `BCUT2D` encoding for `SubstanceParameter`
 - Stored benchmarking results now include the Python environment and version
 - `qPSTD` acquisition function
+- `SHAPInsight` now supports the `waterfall` plot type
 - `ContinuousCardinalityConstraint` is now compatible with `BotorchRecommender`
 - A `MinimumCardinalityViolatedWarning` is triggered when minimum cardinality
   constraints are violated
@@ -26,6 +27,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Acquisition function indicator `is_mc` has been removed in favor of new indicators 
   `supports_batching` and `supports_pending_experiments`
+- `SHAPInsight` now allows explanation input that has additional columns compared to 
+  the background data (will be ignored)
 
 ### Fixed
 - Incorrect optimization direction with `PSTD` with a single minimization target
@@ -33,6 +36,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.12.2] - 2025-01-31
 ### Changed
 - More robust settings for the GP fitting
+- The `beta` parameter of `UCB` and `qUCB` can now also take negative values
 
 ## [0.12.1] - 2025-01-29
 ### Changed